+++ app-react/package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
-    "react-router": "3.0.3"
+    "react-router": "^4.0.0"
   },
   "devDependencies": {
     "react-scripts": "0.9.5"
