+++ app-react/public/css/global.css
@@ -120,6 +120,12 @@ a.nav-item {
 }
 
 /*start categories section*/
+
+.categorySection {
+    padding-top: 70px;
+    padding-bottom: 100px;
+}
+
 .categories {
     text-align: center;
 }
@@ -128,6 +134,21 @@ a.nav-item {
     text-decoration: none;
 }
 
-/*.categoryImage {
+/*Start footer*/
+footer {
+    padding-bottom: 40px;
+    color: #FD41A0;
+    font-size: 50px;
+    text-align: center;
+    font-family: 'Raleway', sans-serif;
+}
 
-}*/
+footer a {
+    color: #FD41A0;
+    /*font-size: 50px;*/
+}
+
+footer a:hover {
+    color: #FD41A0;
+    text-decoration: none;
+}
