+++ app-react/src/components/Category.js
@@ -8,7 +8,7 @@ class Category extends Component {
 
     // You can read incoming props, like categoryTitle="", using this.props.categoryTitle ...wrapping that in {} will output it to the productImage
     render() {
-        return <div className="column category is-one-quarter-desktop">
+        return <div className="column is-one-quarter-desktop">
             <div className="section" style= {{backgroundImage: 'url(' + this.props.categoryImage + ')'}}>
                 <a href={this.props.link} className="catLink">
                     <h3 className="categoryTitle">{this.props.categoryTitle}</h3><br/>
