+++ app-react/src/components/Footer.js
@@ -1,0 +1,20 @@
+import React, { Component } from 'react';
+
+class Footer extends Component {
+  render() {
+    return (
+      <div>
+        <footer class="footer">
+          <div class="container">
+            <div class="content has-text-centered">
+              <a href="">contact</a> | <a href="">about</a>
+            </div>
+          </div>
+        </footer>
+
+      </div>
+    );
+  }
+}
+
+export default Footer;
