+++ app-react/src/components/Product.js
