+++ app-react/src/components/Products.js
@@ -3,7 +3,10 @@ import React, { Component } from 'react';
 class Products extends Component {
   render() {
     return (
-    );
+      // categories . . . no background images
+      // hard code four or so products.
+      <p>products</p>
+    )
   }
 }
