+++ app-react/src/index.js
@@ -1,8 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Home from './Home';
+import Home from './components/Home';
+import Products from './components/Products'
+import { Router, Route, browserHistory, IndexRoute } from 'react-router'
 
 ReactDOM.render(
-  <Home />,
+<Router history={browserHistory}>
+    <Route path="/" component={Home}>
+        <IndexRoute component={Home} />
+        <Route path="products/:productId" component={Products} />
+    </Route>
+  </Router>,
+
   document.getElementById('root')
 );
